make login a component
