+++ app/components/login/template.html
@@ -1,5 +1,5 @@
 <!-- Content Start -->
-<div ng-controller="LoginCtrl" class="container ph-xl-margin-bottom">
+<div class="container ph-xl-margin-bottom">
   <h2 id="login">Log in</h2>
 
   <h4 class="ph-heading-light ph-md-margin-bottom">
